Separation started
