+++ src/main/java/mainBot/MessageProcessor.java
@@ -3,6 +3,9 @@ package mainBot;
 import database.Database;
 import database.models.Connection;
 import database.models.User;
+import mainBot.commandHandlers.CommandHandler;
+import mainBot.commandHandlers.EditHandler;
+import mainBot.commandHandlers.FillingHandler;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,25 +17,14 @@ public class MessageProcessor {
      */
     private final Database database;
     private final Notificator notificator = new Notificator();
+    private final CommandHandler caseCommand;
+    private final FillingHandler caseProfileFill;
+    private final EditHandler caseProfileEdit;
     public MessageProcessor(Database database){
         this.database = database;
-    }
-    /**
-     * Simple help method
-     * @return a description of the commands
-     */
-    private String giveHelp(){
-        return """
-               Вот, что я умею:\s
-                /help - вывод описания всех команд\s
-                /myProfile - посмотреть данные своей анкеты\s
-                /match - поиск собеседника\s
-                /myMatches - посмотреть анкету уже предложенных пользователей\s
-                /pending - посмотреть людей, ожидающих твоего ответа\s
-                /changeProfile - удалить текущую анкету и заполнить новую\s
-                /editProfile - изменить одно из полей анкеты\s
-                /deleteProfile - полностью удалить профиль
-               """;
+        this.caseCommand = new CommandHandler(database);
+        this.caseProfileFill = new FillingHandler(database, stateFSM);
+        this.caseProfileEdit = new EditHandler(database, stateFSM);
     }
     public List<String> getIdList(User sender){
         List<String> idList = new ArrayList<>();
@@ -66,228 +58,6 @@ public class MessageProcessor {
                 reply[14+i] = database.getUser(idList.get(i+page*10)).getPhotoID();
             }else {
                 break;
-            }
-        }
-    }
-    /**
-     * Command handler.
-     * Changes state depending on command.
-     * If message does not start with '/' or is not a supported command - offers to watch command list
-     * @param id string representation of user id
-     * @param message user message
-     * @param sender instance of {@link User} class representing a sender
-     * @param reply array of strings with size of 12, where every string is a separate message
-     */
-    private void caseCommand(String id, String message, User sender, String[] reply){
-        if (database.getUser(id).getExpectedCity() == null){
-            message = "/start";
-        }
-        if (!(message.charAt(0) == '/')){
-            reply[0] = "Что-то я тебя не понимаю, если не знаешь что я умею - введи /help";
-            return;
-        }
-        switch (message){
-            default:
-                reply[0] = "Такой команды нет, введи /help, чтобы увидеть список всех команд";
-                break;
-            case "/start":
-                if (database.getUser(id).getExpectedCity() != null){
-                    reply[0] = giveHelp();
-                    break;
-                }
-                reply[0] = "Привет! Ты попал на MechMatch - место, где тебе помогут найти твою вторую половинку или просто хорошего друга :)  ";
-                reply[1] = """
-                        Перед началом хочется тебя предупредить, что бот никак не идентифицирует пользователя по документам, поэтому будь осторожен!\s
-                        Отправь любое сообщение, чтобы подтвердить прочтение предупреждения.
-                        """;
-                sender.setGlobalState(GlobalState.PROFILE_FILL);
-                sender.setLocalState(LocalState.START);
-                break;
-            case "/help":
-                reply[0] = giveHelp();
-                break;
-            case "/changeProfile":
-                database.eraseProfileData(id);
-                database.deleteFromFPL(id);
-                sender.setProfileFilled(false);
-                reply[0] = "Сейчас тебе придется пройти процедуру заполнения анкеты заново. Напиши что-нибудь, если готов.";
-                sender.setGlobalState(GlobalState.PROFILE_FILL);
-                sender.setLocalState(LocalState.START);
-                break;
-            case "/editProfile":
-                database.deleteFromFPL(id);
-                sender.setProfileFilled(false);
-                reply[0] = "Что хочешь изменить?";
-                reply[1] = "Вот список полей доступных для изменения:" +
-                        " \n1 - Имя(" + sender.getName() +
-                        ")\n2 - Возраст(" + sender.getAge() +
-                        ")\n3 - Пол(" + sender.getSex() +
-                        ")\n4 - Город(" + sender.getCity() +
-                        ")\n5 - Информация о себе(" + sender.getInformation() +
-                        ")\n6 - Нижний порог возраста собеседника(" + sender.getMinExpectedAge() +
-                        ")\n7 - Верхний порог возраста собеседника(" + sender.getMaxExpectedAge() +
-                        ")\n8 - Пол собеседника(" + sender.getExpectedSex() +
-                        ")\n9 - Город собеседника(" + sender.getExpectedCity() +
-                        ")\n10 - Фото";
-                reply[13] = sender.getPhotoID();
-                sender.setGlobalState(GlobalState.PROFILE_EDIT);
-                sender.setLocalState(LocalState.START);
-                break;
-            case "/match":
-                User friend;
-                List<String> fpl = database.getFilledProfilesList(id);
-                int tmpNum = 0;
-                reply[0] = "Не нашлось никого, кто соответствует твоей уникальности ;(";
-                while (tmpNum < fpl.size()) {
-                    friend = database.getUser(fpl.get(tmpNum));
-                    boolean senderSexMatch = (sender.getExpectedSex().equals("без разницы")) || (friend.getSex().equals(sender.getExpectedSex()));
-                    boolean friendSexMatch = (friend.getExpectedSex().equals("без разницы")) || (sender.getSex().equals(friend.getExpectedSex()));
-                    boolean senderCityMatch = (sender.getExpectedCity().equals("любой")) || (friend.getCity().equals(sender.getExpectedCity()));
-                    boolean friendCityMatch = (friend.getExpectedCity().equals("любой")) || (sender.getCity().equals(friend.getExpectedCity()));
-                    boolean senderAgeMatch = (friend.getAge() <= sender.getMaxExpectedAge()) && (friend.getAge() >= sender.getMinExpectedAge());
-                    boolean friendAgeMatch = (sender.getAge() <= friend.getMaxExpectedAge()) && (sender.getAge() >= friend.getMinExpectedAge());
-                    List<String> friendDislikes = new ArrayList<>();
-                    for (Integer i: database.getDislikesOf(friend.getId())){
-                        friendDislikes.add(database.getConnection(i).getFriendID());
-                    }
-                    if (senderSexMatch && senderCityMatch && senderAgeMatch &&
-                            friendSexMatch && friendCityMatch && friendAgeMatch &&
-                            (!database.getAllConnectedUserIds(id).contains(friend.getId())) &&
-                            (!friendDislikes.contains(id))) {
-                        reply[0] = database.profileData(friend.getId());
-                        reply[1] = "Напиши, понравился ли тебе пользователь(да/нет).";
-                        reply[12] = database.getUser(fpl.get(tmpNum)).getPhotoID();
-                        sender.setSuggestedFriendID(friend.getId());
-                        sender.setGlobalState(GlobalState.MATCHING);
-                        break;
-                    }
-                    tmpNum++;
-                }
-                break;
-            case "/myProfile":
-                reply[0] = database.profileData(id);
-                reply[12] = sender.getPhotoID();
-                break;
-            case "/myMatches":
-                if (database.getAllConnectionsWith(id).isEmpty()){
-                    reply[0] = "Просмотренных профилей пока что нет ;(\nПопробуй ввести /match";
-                    return;
-                }
-                reply[0] = "Какой список профилей вывести(лайки/дизлайки)?";
-                reply[1] = "Также из этих списков можно будет удалить профиль по номеру на странице. " +
-                        "А если удаление не требуется, то просто напиши \"выйти\"";
-                sender.setGlobalState(GlobalState.MATCHES);
-                sender.setLocalState(LocalState.CHOICE);
-                break;
-            case "/deleteProfile":
-                reply[0] = "Ты уверен, что хочешь этого? Все твои данные удалятся, в том числе и список понравившихся тебе людей!";
-                reply[1] = "Если ты действительно этого хочешь, то введи свое имя пользователя(то что с собачкой)";
-                sender.setGlobalState(GlobalState.PROFILE_EDIT);
-                sender.setLocalState(LocalState.DELETE);
-                sender.setProfileFilled(false);
-                break;
-            case "/pending":
-                if(database.getPendingOf(id).isEmpty()){
-                    reply[0] = "Нет профилей, ожидающих твоего ответа.";
-                    return;
-                }
-                reply[0] = database.profileData(database.getConnection(database.getPendingOf(id).get(0)).getFriendID());
-                reply[1] = "Напиши, хочешь ли ты начать общение с эти человеком(да/нет)?.";
-                reply[12] = database.getUser(database.getConnection(database.getPendingOf(id).get(0)).getFriendID()).getPhotoID();
-                sender.setGlobalState(GlobalState.PENDING);
-                break;
-        }
-    }
-
-    /**
-     * Profile filling procedure handler.
-     * Fills the exact profile data fields of given user depending on its state
-     * @param id string representation of user id
-     * @param message user message
-     * @param sender instance of {@link User} class representing a sender
-     * @param reply array of strings with size of 12, where every string is a separate message
-     */
-    private void caseProfileFill(String id, String message, User sender, String[] reply) {
-        switch (sender.getLocalState()) {
-            case START:
-                reply[0] = "А теперь перейдем к заполнению анкеты.";
-                reply[1] = "Введи свое имя";
-                sender.setLocalState(LocalState.NAME);
-                break;
-            case FINISH:
-                if (message.equalsIgnoreCase("да")) {
-                    database.addToFPL(id);
-                    sender.setProfileFilled(true);
-                    reply[0] = "Отлично, теперь можно переходить к использованию.";
-                    reply[1] = giveHelp();
-                    sender.setGlobalState(GlobalState.COMMAND);
-                } else if (message.equalsIgnoreCase("нет")) {
-                    reply[0] = "Что хочешь изменить?";
-                    reply[1] = "Вот список полей доступных для изменения:" +
-                            " \n1 - Имя(" + sender.getName() +
-                            ")\n2 - Возраст(" + sender.getAge() +
-                            ")\n3 - Пол(" + sender.getSex() +
-                            ")\n4 - Город(" + sender.getCity() +
-                            ")\n5 - Информация о себе(" + sender.getInformation() +
-                            ")\n6 - Нижний порог возраста собеседника(" + sender.getMinExpectedAge() +
-                            ")\n7 - Верхний порог возраста собеседника(" + sender.getMaxExpectedAge() +
-                            ")\n8 - Пол собеседника(" + sender.getExpectedSex() +
-                            ")\n9 - Город собеседника(" + sender.getExpectedCity() +
-                            ")\n10 - Фото";
-                    sender.setGlobalState(GlobalState.PROFILE_EDIT);
-                    sender.setLocalState(LocalState.START);
-                    break;
-                } else {
-                    reply[0] = "Пожалуйста, напиши либо да, либо нет";
-                }
-                break;
-            default:
-                if (sender.setField(message)) {
-                    reply[0] = stateFSM.getRightReplies().get(sender.getLocalState());
-                    sender.setLocalState(stateFSM.getNextDict().get(sender.getLocalState()));
-                } else {
-                    reply[0] = stateFSM.getWrongReplies().get(sender.getLocalState());
-                }
-                break;
-        }
-    }
-    /**
-     * Profile editing procedure handler.
-     * Fills the exact profile data field of given user depending on the user's choice
-     * @param id string representation of user id
-     * @param message user message
-     * @param sender instance of {@link User} class representing a sender
-     * @param reply array of strings with size of 12, where every string is a separate message
-     */
-    private void caseProfileEdit(String id, String message, User sender, String[] reply){
-        if (sender.getLocalState().equals(LocalState.START)) {
-            if (!stateFSM.getStateDict().containsKey(message)) {
-                reply[0] = "Напиши либо цифру соответствующую полю, либо название поля.";
-                return;
-            }
-            sender.setLocalState(stateFSM.getStateDict().get(message));
-            reply[0] = stateFSM.getEditReplies().get(sender.getLocalState());
-        }
-        else if (sender.getLocalState() == LocalState.DELETE){
-            if (message.equals(sender.getUsername()) || message.equals("@" + sender.getUsername())){
-                reply[0] = "Профиль успешно удален.";
-                database.deleteUser(id);
-                database.deleteAllConnectionsWith(id);
-                return;
-            }
-            reply[0] = "Введено неверное значение, процедура удаления прекращена.";
-            sender.setGlobalState(GlobalState.COMMAND);
-            database.addToFPL(id);
-            sender.setProfileFilled(true);
-        }
-        else {
-            if (sender.setField(message)) {
-                reply[0] = "Изменение внесено.";
-                sender.setGlobalState(GlobalState.COMMAND);
-                database.addToFPL(id);
-                sender.setProfileFilled(true);
-            } else {
-                reply[0] = stateFSM.getWrongReplies().get(sender.getLocalState());
             }
         }
     }
@@ -485,9 +255,9 @@ public class MessageProcessor {
         }
         User sender = database.getUser(id);
         switch (sender.getGlobalState()){
-            case COMMAND -> caseCommand(id, message, sender, reply);
-            case PROFILE_FILL -> caseProfileFill(id, message, sender, reply);
-            case PROFILE_EDIT -> caseProfileEdit(id, message, sender, reply);
+            case COMMAND -> caseCommand.handleMessage(sender, reply, message);
+            case PROFILE_FILL -> caseProfileFill.handleMessage(sender, reply, message);
+            case PROFILE_EDIT -> caseProfileEdit.handleMessage(sender, reply, message);
             case MATCHES -> caseMatches(message, sender, reply);
             case MATCHING -> caseMatching(id, message, sender, reply);
             case PENDING -> casePending(id, message, sender, reply);
