+++ src/main/java/mainBot/commandHandlers/EditHandler.java
@@ -1,0 +1,52 @@
+package mainBot.commandHandlers;
+
+import database.Database;
+import database.models.User;
+import mainBot.GlobalState;
+import mainBot.LocalState;
+import mainBot.StateFSM;
+
+/**
+ * Profile editing procedure handler.
+ * Fills the exact profile data field of given user depending on the user's choice
+ */
+public class EditHandler implements Handler{
+    private final Database database;
+    private final StateFSM stateFSM;
+    public EditHandler(Database m_database, StateFSM m_stateFSM){
+        this.database = m_database;
+        this.stateFSM = m_stateFSM;
+    }
+    public void handleMessage(User sender, String[] reply, String message) {
+        if (sender.getLocalState().equals(LocalState.START)) {
+            if (!stateFSM.getStateDict().containsKey(message)) {
+                reply[0] = "Напиши либо цифру соответствующую полю, либо название поля.";
+                return;
+            }
+            sender.setLocalState(stateFSM.getStateDict().get(message));
+            reply[0] = stateFSM.getEditReplies().get(sender.getLocalState());
+        }
+        else if (sender.getLocalState() == LocalState.DELETE){
+            if (message.equals(sender.getUsername()) || message.equals("@" + sender.getUsername())){
+                reply[0] = "Профиль успешно удален.";
+                database.deleteUser(sender.getId());
+                database.deleteAllConnectionsWith(sender.getId());
+                return;
+            }
+            reply[0] = "Введено неверное значение, процедура удаления прекращена.";
+            sender.setGlobalState(GlobalState.COMMAND);
+            database.addToFPL(sender.getId());
+            sender.setProfileFilled(true);
+        }
+        else {
+            if (sender.setField(message)) {
+                reply[0] = "Изменение внесено.";
+                sender.setGlobalState(GlobalState.COMMAND);
+                database.addToFPL(sender.getId());
+                sender.setProfileFilled(true);
+            } else {
+                reply[0] = stateFSM.getWrongReplies().get(sender.getLocalState());
+            }
+        }
+    }
+}
